rename: move api package into remote package
