+++ app/src/main/java/com/ottfstudio/quizdroid/data/remote/DefaultQuizDataSource.kt
@@ -1,8 +1,8 @@
 package com.ottfstudio.quizdroid.data.remote
 
-import com.ottfstudio.quizdroid.data.api.QuizService
 import com.ottfstudio.quizdroid.data.datasource.QuizDataSource
 import com.ottfstudio.quizdroid.data.model.QuizResponse
+import com.ottfstudio.quizdroid.data.remote.api.QuizService
 import javax.inject.Inject
 import javax.inject.Singleton
